refactor(tasks): move task actions into data hook
